% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KneeOsteoarthritis.R
\docType{data}
\name{KneeOsteoarthritis}
\alias{KneeOsteoarthritis}
\title{Data of 4,796 individuals with or without knee osteoarthritis}
\format{A dataset with 4,796 rows and 2 columns:
\describe{
  \item{ID}{Unique identifier variable for each individual}
  \item{KNEEOA}{Diagnosis of knee osteoarthritis, coded as 0 (not affected) and 1 (affected)}
  ...

}}
\source{
\url{https://nda.nih.gov/oai/}
}
\usage{
KneeOsteoarthritis
}
\description{
The dataset `KneeOsteoarthritis` contains data of 4,796 individuals (including the 4,704 individuals of the `HandOsteoarthritis` datasets) with or without knee osteoarthritis.
}
\keyword{datasets}
