% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeitmice.R
\name{mergeitmice}
\alias{mergeitmice}
\title{Merges Imputed Datasets with Dataframes}
\usage{
mergeitmice(datasets, data, by = "ID")
}
\arguments{
\item{datasets}{This argument specifies an object of the \code{mids}, \code{mimids}, or \code{wimids} class.}

\item{data}{This argument specifies a data frame.}

\item{by}{This argument specifies a variable name, present in bot \code{datasets} and \code{data}.}
}
\value{
This function returns an object of the \code{mids}, \code{mimids}, or \code{wimids} class after merging a dataframe with each imputed dataset of the inputted object.
}
\description{
The \code{mergeitmice()} function merges a dataframe with each imputed dataset of the \code{mids}, \code{mimids}, or \code{wimids} class objects based on the variables passed to the function as \code{by}.
}
\details{
This functions can be used instead of the \code{cbind()} function (from the \pkg{mice} package).
}
\examples{
\donttest{
#Please see the package repository <https://github.com/FarhadPishgar/MatchIt.mice> for details.

#Loading and preparing the 'handoa' dataset
data(handoa)
idenoa <- handoa["ID", "HANDOA"]
handoa <- handoa[c("ID", "AGE", "SEX", "BMI", "SMOKING", "HANDUSE", "KNEEOA")]

#Imputing the missing data points in the'handoa' dataset
datasets <- mice(handoa, m = 5, maxit = 10)

#Merging the dataframe, 'idenoa', with each imputed dataset of the 'datasets' object
datasets <- mergeitmice(datasets, idenoa, by = "ID")
}
}
\references{
Stef van Buuren and Karin Groothuis-Oudshoorn (2011). \code{mice}: Multivariate Imputation by Chained Equations in \code{R}. \emph{Journal of Statistical Software}, 45(3): 1-67. \url{https://www.jstatsoft.org/v45/i03/}
}
\seealso{
\code{\link[=matchitmice]{matchitmice}}

\code{\link[=weightitmice]{weightitmice}}

\code{\link[=binditmice]{binditmice}}
}
\author{
Farhad Pishgar
}
\keyword{functions}
